bet-maker/pyproject.toml
[tool.poetry]
name = "bet-maker"
version = "0.1.0"
description = "Making bets!"
authors = ["Maxim Anufriev <anufrieff.maxim@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
aioredis = "1.3.1"
aio-pika = "9.4.2"
alembic = "1.13.2"
asyncpg = "0.29.0"
fastapi = "0.111.1"
aiormq = "6.8.0"
annotated-types = "0.7.0"
anyio = "4.4.0"
async-timeout = "4.0.3"
certifi = "2024.7.4"
click = "8.1.7"
databases = "0.9.0"
dnspython = "2.6.1"
email-validator = "2.2.0"
exceptiongroup = "1.2.2"
fastapi-cache2 = "0.2.2"
fastapi-cli = "0.0.4"
greenlet = "3.0.3"
h11 = "0.14.0"
hiredis = "3.0.0"
httpcore = "1.0.5"
httptools = "0.6.1"
httpx = "0.27.0"
idna = "3.7"
jinja2 = "3.1.4"
mako = "1.3.5"
markdown-it-py = "3.0.0"
markupsafe = "2.1.5"
mdurl = "0.1.2"
multidict = "6.0.5"
mypy-extensions = "1.0.0"
packaging = "24.1"
pamqp = "3.3.0"
pathspec = "0.12.1"
pendulum = "3.0.0"
pika = "1.3.2"
platformdirs = "4.2.2"
psycopg2-binary = "2.9.9"
pydantic = "2.8.2"
pydantic-core = "2.20.1"
pygments = "2.18.0"
python-dateutil = "2.9.0.post0"
python-dotenv = "1.0.1"
python-multipart = "0.0.9"
pyyaml = "6.0.1"
redis = "4.6.0"
rich = "13.7.1"
shellingham = "1.5.4"
six = "1.16.0"
sniffio = "1.3.1"
sqlalchemy = "2.0.31"
starlette = "0.37.2"
time-machine = "2.14.2"
tomli = "2.0.1"
typer = "0.12.3"
typing-extensions = "4.12.2"
tzdata = "2024.1"
uvicorn = "0.30.4"
watchfiles = "0.22.0"
websockets = "12.0"
yarl = "1.9.4"
types-redis = "^4.6.0.20241004"

[tool.poetry.group.dev.dependencies]
ruff = "^0.9.2"
pytest = "^8.3.4"
pytest-cov = "^6.0.0"
mypy = "^1.14.1"
black = "^24.10.0"
isort = "^5.13.2"
pre-commit = "^4.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "G004", "D103", "D100", "D101",
    "B904", "D104", "D203", "D212",
    "TRY301", "ANN201", "RUF006",
]

[tool.ruff.format]
quote-style = "double"

[tool.mypy]
python_version = "3.12"
files = "app"
strict = true
ignore_missing_imports = true
show_error_codes = true
pretty = true
warn_unused_configs = true
